toolkit/components/passwordmgr/test/mochitest/test_autocomplete_password_open.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test password field autocomplete footer with and without logins</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<script>
var chromeScript = runChecksAfterCommonInit();

var setupScript = runInParent(function setup() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  var nsLoginInfo = Components.Constructor("@mozilla.org/login-manager/loginInfo;1",
                                           Ci.nsILoginInfo, "init");
  assert.ok(nsLoginInfo != null, "nsLoginInfo constructor");

  var login1 = new nsLoginInfo("https://example.com", "", null,
                               "tempuser1", "temppass1", "uname", "pword");
  var login2 = new nsLoginInfo("https://example.com", "", null,
                               "tempuser2", "temppass2", "uname", "pword");

  // try/catch in case someone runs the tests manually, twice.
  try {
    Services.logins.addLogin(login1);
    Services.logins.addLogin(login2);
  } catch (e) {
    assert.ok(false, "addLogin threw: " + e);
  }
});
</script>
<p id="display"></p>

<div id="content">
  <h1>Sign in</h1>
  <input id="uname"/>
  <input id="pword" type="password"/>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: Test password field autocomplete footer with and without logins **/

let uname = document.getElementById("uname");
let pword = document.getElementById("pword");

add_task(async function test_no_autofill() {
  // Make sure initial form is empty as autofill shouldn't happen in the sandboxed frame.
  checkLoginForm(uname, "", pword, "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

add_task(async function test_two_logins() {
  let shownPromise = promiseACShown();
  uname.focus();
  await shownPromise;

  shownPromise = promiseACShown();
  synthesizeKey("KEY_Tab"); // popup on the password field should open upon focus
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  let expectedMenuItems = [
    "tempuser1",
    "tempuser2",
  ];
  checkAutoCompleteResults(results, expectedMenuItems, "example.com", "Check all menuitems are displayed correctly.");

  checkLoginForm(uname, "", pword, "");
  let removedPromise = promiseStorageChanged(["removeAllLogins"]);
  await LoginManager.removeAllLogins();
  await removedPromise;
});

add_task(async function test_zero_logins() {
  uname.focus();

  let shownPromise = promiseACShown().then(() => ok(false, "Should not have shown"));
  // Popup on the password field should NOT automatically open upon focus when there are no saved logins.
  synthesizeKey("KEY_Tab"); // focus the password field
  SimpleTest.requestFlakyTimeout("Giving a chance for the unexpected popup to show");
  let results = await Promise.race([
    shownPromise,
    new Promise(resolve => setTimeout(resolve, 2000)), // Wait 2s for the popup to appear
  ]);

  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is still closed");

  checkLoginForm(uname, "", pword, "");
  shownPromise = promiseACShown();
  info("arrow down should still open the popup");
  synthesizeKey("KEY_ArrowDown");
  results = await shownPromise;
  checkAutoCompleteResults(results, [], "example.com", "Check only footer is displayed.");
  checkLoginForm(uname, "", pword, "");
});
</script>
</pre>
</body>
</html>
